fix : refactoring code and test case
