+++ src/main/java/com/fesi/mukitlist/api/controller/annotation/CustomAuthenticationPrincipalResolver.java
@@ -12,7 +12,7 @@ import org.springframework.web.method.support.ModelAndViewContainer;
 
 import com.fesi.mukitlist.api.exception.AppException;
 import com.fesi.mukitlist.core.auth.PrincipalDetails;
-import com.fesi.mukitlist.core.auth.User;
+import com.fesi.mukitlist.core.auth.application.User;
 
 public class CustomAuthenticationPrincipalResolver implements HandlerMethodArgumentResolver {
