+++ src/main/java/com/fesi/mukitlist/api/controller/auth/UserController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.fesi.mukitlist.api.controller.auth.request.UserUpdateRequest;
-import com.fesi.mukitlist.domain.service.auth.UserService;
+import com.fesi.mukitlist.domain.service.auth.application.UserService;
 import com.fesi.mukitlist.domain.service.auth.response.UserInfoResponse;
 import com.fesi.mukitlist.core.auth.PrincipalDetails;
 import com.fesi.mukitlist.api.controller.annotation.Authorize;
