+++ src/main/java/com/fesi/mukitlist/api/controller/auth/oauth/kakao/KakaoTokenController.java
@@ -1,9 +1,9 @@
 package com.fesi.mukitlist.api.controller.auth.oauth.kakao;
 
 
-import static com.fesi.mukitlist.core.auth.constant.GrantType.*;
-import static com.fesi.mukitlist.core.auth.constant.TokenType.*;
-import static com.fesi.mukitlist.core.auth.constant.UserType.*;
+import static com.fesi.mukitlist.core.auth.GrantType.*;
+import static com.fesi.mukitlist.core.auth.TokenType.*;
+import static com.fesi.mukitlist.core.auth.application.constant.UserType.*;
 
 import java.util.Map;
 import java.util.Objects;
@@ -29,12 +29,12 @@ import com.fesi.mukitlist.api.repository.TokenRepository;
 import com.fesi.mukitlist.api.repository.UserRepository;
 import com.fesi.mukitlist.core.auth.PrincipalDetails;
 import com.fesi.mukitlist.core.auth.Token;
-import com.fesi.mukitlist.core.auth.User;
-import com.fesi.mukitlist.core.oauth.KakaoUserInfo;
-import com.fesi.mukitlist.core.oauth.OAuth2UserInfo;
+import com.fesi.mukitlist.core.auth.application.User;
+import com.fesi.mukitlist.core.auth.oauth.KakaoUserInfo;
+import com.fesi.mukitlist.core.auth.oauth.OAuth2UserInfo;
 import com.fesi.mukitlist.domain.service.auth.AuthenticationService;
 import com.fesi.mukitlist.domain.service.auth.JwtService;
-import com.fesi.mukitlist.domain.service.auth.UserService;
+import com.fesi.mukitlist.domain.service.auth.application.UserService;
 
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
