+++ src/main/java/com/fesi/mukitlist/core/auth/application/User.java
@@ -1,14 +1,13 @@
-package com.fesi.mukitlist.core.auth;
+package com.fesi.mukitlist.core.auth.application;
 
 import java.time.LocalDateTime;
 
-import org.springframework.data.annotation.CreatedBy;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import com.fesi.mukitlist.api.controller.auth.oauth.kakao.request.KakaoServiceCreateRequest;
-import com.fesi.mukitlist.core.auth.constant.UserType;
+import com.fesi.mukitlist.core.auth.application.constant.UserType;
 import com.fesi.mukitlist.domain.service.auth.request.UserServiceCreateRequest;
 
 import jakarta.persistence.Column;
