+++ src/main/java/com/fesi/mukitlist/core/usergathering/UserGatheringId.java
@@ -1,9 +1,8 @@
 package com.fesi.mukitlist.core.usergathering;
 
-import com.fesi.mukitlist.core.auth.User;
+import com.fesi.mukitlist.core.auth.application.User;
 import com.fesi.mukitlist.core.gathering.Gathering;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
