+++ src/main/java/com/fesi/mukitlist/domain/service/auth/oauth/PrincipalService.java
@@ -1,4 +1,4 @@
-package com.fesi.mukitlist.domain.service.oauth;
+package com.fesi.mukitlist.domain.service.auth.oauth;
 
 import static com.fesi.mukitlist.api.exception.ExceptionCode.*;
 
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 import com.fesi.mukitlist.api.exception.AppException;
 import com.fesi.mukitlist.api.repository.UserRepository;
 import com.fesi.mukitlist.core.auth.PrincipalDetails;
-import com.fesi.mukitlist.core.auth.User;
+import com.fesi.mukitlist.core.auth.application.User;
 
 import lombok.RequiredArgsConstructor;
