+++ src/main/java/com/fesi/mukitlist/domain/service/auth/oauth/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.fesi.mukitlist.api.service.oauth.response;
+package com.fesi.mukitlist.domain.service.auth.oauth.response;
 
 public record LoginResponse(String code, String accessToken) {
 }
