+++ src/main/java/com/fesi/mukitlist/domain/service/auth/request/TokenServiceRequest.java
@@ -1,8 +1,8 @@
 package com.fesi.mukitlist.domain.service.auth.request;
 
-import com.fesi.mukitlist.core.auth.constant.GrantType;
-import com.fesi.mukitlist.core.auth.constant.TokenType;
-import com.fesi.mukitlist.core.auth.User;
+import com.fesi.mukitlist.core.auth.GrantType;
+import com.fesi.mukitlist.core.auth.TokenType;
+import com.fesi.mukitlist.core.auth.application.User;
 import lombok.Builder;
 
 @Builder
