+++ src/main/java/com/fesi/mukitlist/domain/service/review/request/ReviewServiceCreateRequest.java
@@ -1,6 +1,6 @@
 package com.fesi.mukitlist.domain.service.review.request;
 
-import com.fesi.mukitlist.core.auth.User;
+import com.fesi.mukitlist.core.auth.application.User;
 import com.fesi.mukitlist.core.gathering.Gathering;
 import com.fesi.mukitlist.core.Review;
